extract ItemSubCommand, GiveSubCommand, DebugSubCommand, DownloadSubCommand,VersionSubCommand, GlobalPointSubCommand, PointSubCommand, GlobalTagSubCommand, TagSubCommand, VariableObjectiveSubCommand, ObjectiveSubCommand, JournalSubCommand, and ActionSubCommand from QuestCommand
